Split websocket.js to client.js and manager.js
